feat: jwt signing keys front-and-center
