+++ apps/studio/components/layouts/JWTKeys/JWTKeysLayout.tsx
@@ -3,26 +3,20 @@ import { ScaffoldContainer } from 'components/layouts/Scaffold'
 import { PropsWithChildren } from 'react'
 
 import { useParams } from 'common'
-import { useIsFeatureEnabled } from 'hooks/misc/useIsFeatureEnabled'
 
 const JWTKeysLayout = ({ children }: PropsWithChildren) => {
   const { ref: projectRef } = useParams()
-  const { projectSettingsLegacyJwtKeys } = useIsFeatureEnabled(['project_settings:legacy_jwt_keys'])
 
   const navigationItems = [
-    ...(projectSettingsLegacyJwtKeys
-      ? [
-          {
-            label: 'Legacy JWT Secret',
-            href: `/project/${projectRef}/settings/jwt`,
-            id: 'legacy-jwt-keys',
-          },
-        ]
-      : []),
     {
       label: 'JWT Signing Keys',
-      href: `/project/${projectRef}/settings/jwt/signing-keys`,
+      href: `/project/${projectRef}/settings/jwt`,
       id: 'signing-keys',
+    },
+    {
+      label: 'Legacy JWT Secret',
+      href: `/project/${projectRef}/settings/jwt/legacy`,
+      id: 'legacy-jwt-keys',
     },
   ]
